add get-last-rec-from-host
This script will show when the last record from each host was
stored in Elasticsearch, along with the count of records from
that host.  It performs an aggregate search by the `"hostname"`
field, which is usually the Kubernetes nodename.

The output looks like this:
```
green : mynode.compute.internal diff 0:00:04.773011 records 50695
```
This means that the last record from node `mynode.compute.internal`
has a `@timestamp` value of 4.77 seconds ago, and that 50695
records have been received in the time interval of the search.

By default, it will go back 3 hours.  Use the env. var. `OLDEST`
to change the time interval e.g.
```
OLDEST=24h ./get-last-rec-from-host
```
to view the past day, or
```
OLDEST=15m ./get-last-rec-from-host
```
to view the last 15 minutes